Do not allow prefixing zero on fuzzy hash strings
A fuzzy hash with prefixing '0' will never be generated by ssdeep and
we don't have a good reason to support it.

By prohibiting it explicitly, it improves canonicality.

That means, if a fuzzy hash string does not contain "," and no normalization
occurs, converting to a fuzzy hash object and then back to a string always
results in the original string (no change except trailing "," and later
characters).