Add light-weight prefiltering methods
`is_comparison_candidate` and its variants are new methods for light-weight
prefiltering methods.

This operation is useful to divide a set of *unique* (normalized) fuzzy
hashes into smaller distinct sets.  The similarity score can be non-zero if
and only if two fuzzy hashes belong to the same set.

Note however, this function (and its variants) can return `false` if `self`
and `other` are equivalent (the base fuzzy hash object of `self` and `other`
are the same and their similarity score is 100).

Because of this, we have to use a set of *unique* fuzzy hash values
on the use case above to prevent false-negative matches.