Add invariants to block_hash_[12]()
The author originally intended `block_hash_1()` and `block_hash_2()` methods
to be raw user access methods and not to be particularly fast.  That's why
they did not have any invariants (despite that they had slice accesses).

However, this crate utilized them in frequently called core functions and
the author found that it slightly slowed the program.

This commit adds invariants for them for maximum performance.  It eludes
excess branches and improves the performance slightly.