Add crate global `MAX_LEN_IN_STR`
We often use the maximum length of the valid fuzzy hash possible but there
are only `Long{,Raw}FuzzyHash::MAX_LEN_IN_STR` for this purpose.

It would not be clear enough and this commit adds the new crate global
`MAX_LEN_IN_STR` constant.  It would also be a substitute for
`FUZZY_MAX_RESULT` in the original libfuzzy.

This is a backport from the version 0.2.x development branch.