doc: Clarify generator sample much more
Using `core::iter::repeat` for just one byte is too much and we haven't
introduced the `update_by_byte()` method.

This commit improves the clarity of a generator example.

As a side effect, it fixes a problem which the same examples both in lib.rs
and generate.rs were not synchronized.