Rename `as_ref_normalized` to `as_normalized`
To follow conventional Rust API naming, `as_ref_normalized` is too
redundant.  `as_normalized` is simpler and follows the same naming scheme
as `str::as_bytes`.

This commit renames the method but adds the `as_ref_normalized` method
as a mirror to `as_normalized`.  This is going to be removed on the next
major release.