tests: Make `test_blockhash_contents_no_sequences` crate private
This commit makes this method from (implicit) private to `pub(crate)`
(crate private) to make this function available to other modules
in this crate.

This is a subset of all test cases (in fact, we cannot test normalization
with this) but deterministic and easy to modify (for required tests).

To make it clear, it adds a comment to describe the block hashes generated.