Allow old lint names
`clippy::derive_hash_xor_eq` is the problem.  While it was default error
on our current MSRV (1.56), it is now renamed and the meaning of the lint
has changed (so that our code is no longer affected by the renamed lint).

Related (rust-clippy):
Allow implementing `Hash` with derived `PartialEq` (`derive_hash_xor_eq`)
<https://github.com/rust-lang/rust-clippy/pull/10184>

As a workaround, it disables `renamed_and_removed_lints` lint unless in
the maintaince mode.