Add new Clippy lint allow and add a workaround
`impl core::convert::From<FuzzyHashData<S1, S2, true>` for
`FuzzyHashCompareTarget` required allowing `clippy::needless_borrow`
but this is captured by `clippy::needless_borrows_for_generic_args`
instead in Rust 1.74.  This commit adds it.

However, it revealed that adding `#![allow(clippy::...)]` was not enough.
Due to bugs on lint handling on Rust -1.62, we need to unconditionally
allow `unknown_lints` in the method, implementation or the module
(adding to lib.rs does not work) to make Rust -1.62 happy.  This commit
adds `#[allow(unknown_lints)]` to the method.

Once MSRV 1.63 is acceptable, we need to remove this
`#[allow(unknown_lints)]`.