Optimize `FuzzyHashData::compare{,_unequal}` for sure
Commit 69054377a72d (Optimize non-batching safe `compare` function) turned
out to be slower than before on many cases and the author noticed a simple
mistake: to reduce costly initialization of `FuzzyHashCompareTarget`, we
have to expand the contents of the `FuzzyHashCompareTarget` method to call.

This commit effectively expands the contents of corresponding methods
on `FuzzyHashCompareTarget` before constructing it and makes batch safe
comparison significantly faster.

There's a big room for improvements but the author implemented the
optimization which doesn't (effectively) need a new type.