Optimization to comparison with two fuzzy hashes
`FuzzyHashData::compare()` method provides easy access to the comparison
function and at the same time we can skip the parser (allowing fast and
easy cluster matching).  Although fully utilizing `FuzzyHashCompareTarget`
is one of the fastest ways to do batch clustering, the author cannot ignore
the performance of relatively easy functions.

This commit adds "half-baked `FuzzyHashCompareTarget`" optimization.
While previously, we constructed `FuzzyHashCompareTarget` with two block
hashes but if the block sizes are different, we don't need initialization
of at least one of the block hashes (since it's unused while comparison).

Previously, `BlockHashPositionArray` struct was a mere demonstration to
represent **a** block hash using the position array representation.
However, utilizing this struct when the block sizes are different but near,
we can initialize the position array for one of the block hashes, not two.

On the author's environment, brute force comparison inside a set with
various block sizes (using `FuzzyHashData::compare()`) showed the
performance improvement of 16-31% (varies by the data set, the `unsafe`
feature, and the operand order (`a.compare(b)` or `b.compare(a)`)).