Remove an invariant on `parse_block_hash_from_bytes`
Since LLVM can automatically infer the boundary checking is not required
(even without this invariant), remove this invariant to let the optimizer
do the job.

As a result, in the latest Nightly Rust + LLVM (with high optimization),
only two boundary checks will occur when we parse a fuzzy hash from the
string (`from_bytes_with_last_index`; even if the `unsafe` feature
is turned off).

This is one of the reasons why enabling the `unsafe` feature won't improve
the parser performance.