Optimize naïve `edit_distn`
We used two vectors to represent last two rows of the table (the previously
updated one and the one we are now updating) but redundant.

This commit optimizes the algorithm to use only one row and three
additional variables (two of them are required to represent previous row and
one of them is for purely readability and will be optimized out).

The speed-up mainly comes from the halved vector allocation cost.

As a result, this function became partially similar to that of the
`edit-distance` crate but no credit is added since there's pretty few
choices to optimize the original code (thus considered non-copyrightable).