Fix roles of two costs on naïve `edit_distn`
Although that the algorithm itself were fine, two variables, `cost_a` and
`cost_d` are named by wrong assumptions.  Previously, `cost_a` represented
the cost for a character deletion and `cost_d` represented the cost for a
character addition.  This is the opposite of what we expect.

This commit fixes variable names.