Tsukasa's simplified LCS distance algorithm
Inspired by Myers (1999) (doi:10.1145/316542.316550), Tsukasa OI (the author
of this crate) created the new algorithm optimized for the LCS distance,
reducing the number of hot spot instructions and the number of variables
to keep track.

The validness of this algorithm (naïve DP to bit-parallel conversion) is
formally-verified using Z3 and this commit also adds the supplemental formal
verification code (along with the formal verification of the old LCS
distance algorithm).

Despite that it significantly speeds up the edit distance code, the actual
speedup in the comparison is not high.  This is because the main bottleneck
is not the computation of the edit distance (rather, finding a common
substring in two block hashes takes around 80% of the comparison time).