Version 0.2.8: "New Year Comes with a Great Advancement"
While many people are celebrating the new year (and/or evacuating due to
a large earthquake disrupting the Japanese new year), I was in the middle
of "accidental discovery".

Although the performance improvement (as overall) is not high enough, I
implemented the new simplified algorithm to compute the LCS distance
(that is formally-verified).  I decided to formally prove previous LCS
distance code and... fortunately, I successfully proved that.

That is going to be a great milestone for reliability of this crate,
along with the upstream ssdeep.

As usual, it paves the way forward for upcoming versions 0.3 and 0.4.

*   Fixes (does not change the behavior)
    *   Two variable names in `edit_distn` were named incorrectly
        (functionally fine though).
*   Optimization
    *   Completely new simplified algorithm to compute edit distance
        (LCS distance) between two strings (`edit_distance` family).
        The author measured about 70% speedup on the LCS distance code
        alone but overall comparison performance improvement will not be
        larger than 5%.
    *   Half vector allocation on `edit_distn` test edit distance
        (LCS distance) computation function to speed up the testing.
*   Reliability
    *   Formal Verification of the LCS computation code
        (the new one and the old one).
*   Copyright
    *   Update copyright year to 2024 on some files.