Version 0.2.9: "Over and Above"
My "new" algorithm is retired in a day.  Not because it's wrong
but because I was told that the far better algorithm exists and ...
found that it was true!

Great thanks to Hajime Suzuki (ocxtal) for suggesting the algorithm!

Anyway, reliability is a prime factor.  The author is adding more formal
verification code instead of relying on the fuzzing.

As usual, it paves the way forward for upcoming versions 0.3 and 0.4.

*   Performance
    *   The LCS distance is now calculated from the length of the longest
        common subsequence between two strings.
        `edit_distance` family method got about 3 times faster thanks to
        the LCS length algorithm by Hyyrö (2004).
*   Reliability
    *   Editorial fixes and minor generalization to the existing prover.
    *   New formal verification to the rolling hash.
*   Tests
    *   Removed slow fuzzing code for the rolling hash
        (since main properties are now formally verified).
    *   Enhancements and simplifications.