Version 0.2.10: "Minimax"
While preserving the reliability, the author is going to minimize the
tests that are already tested by other test cases.  The slow test cases
are wasting much time and we already know that we are testing something
unnecessary if we trust previous tests ran.

Formal verification is one of the ways to minimize the tests
but not the only one.

*   Reliability
    *   New formal verification to Hyyrö (2004).
*   Documentation
    *   Minor fixes (including editorial ones).
*   Tests
    *   Added rolling hash internal inspection code.
    *   Removed fast fuzzing code for the rolling hash.
    *   Catched up with the latest Clippy.