generator: Simplification using `cfg_if`
The author didn't utilize `cfg_if` much because the author did not know
that `cfg_if!` "block" does not actually make a block.  For instance, we can
use variables defined in the `cfg_if!` block outside it.

Because the generator is not so simple, simplifying the code is vital.
By fully utilizing the capabilities of `cfg_if`, we could greatly reduce
the complexity of the generator macro.