Optimize normalization handlings (`u32` to `usize`)
The reason we initially used `u32` for indices is, at the time the author
made an optimized engine for ssdeep, using 32-bit integer was slightly
faster.  But this is not a semantically good practice in general.

Additionally, recent benchmark suggested that replacing such `u32` to
`usize`, actually improves the performance (around 5%) on recent x86_64
processors.

So, we don't have any reason to preserve `u32` in it.
This commit replaces such uses of `u32` in `hash::algorithms`:

*   `normalize_block_hash_in_place()`
*   `is_normalized()`
*   `parse_block_hash_from_bytes()`