Optimize `FuzzyHashDualData::from_bytes()` with direct construction
Before this commit, a `FuzzyHashDualData` object is constructed by first
constructing a raw fuzzy hash object and then compressing it.  It induced
a significant loss (despite that string to dual fuzzy hash construction)
can be a common component of fast fuzzy hash processing programs.

This commit introduces direct construction of the `FuzzyHashDualData`
object from the string (or byte string).  This is as fast as feeding a raw
fuzzy object into `FuzzyHashDualData::from()` (a bit slower than that but
because full parser does the job, that's understandable).

The author measured following micro-benchmark results (using Criterion.rs)
on the development environment:

*    71.84ns: String to raw fuzzy hash
*   130.15ns: Raw fuzzy hash to dual fuzzy hash
*   201.99ns: (calculated) total of the above two (old)

*   170.68ns: String to dual fuzzy hash via raw fuzzy hash (old)
    Possibly because optimization and better memory locality, it's faster
    than calculated sum of two operations.
*   139.84ns: String to dual fuzzy hash (new!)