hash: Merge inherent normalization implementations
By making the `normalize_block_hash_in_place()` internal function
to perform normalization only if the expected original form is raw,
this commit merges two distinct implementations of `FuzzyHashData` methods:

*   `normalize()`
*   `clone_normalized()`

It improves the readability of the manual and more importantly, reduces
the necessity to split implementations (even if we covered both `true`
and `false` normalization parameters, we could not use methods above
in a generic function accepting both normalized and raw forms before
this commit).

We consider this change non-breaking and confirmed that (on higher
optimization levels) its performance impact is negligible.