hash: Merge inherent normalization verification
By making the `algorithms::is_normalized()` internal function parameterized
(by new `is_normalized_internal()` function) to check the normalization
only if certain condition is satisfied, this commit merges two distinct
implementations of `FuzzyHashData` method:

*   `is_normalized()`

It improves the readability of the manual and more importantly, reduces
the necessity to split implementations (even if we covered both `true`
and `false` normalization parameters, we could not use the method above
in a generic function accepting both normalized and raw forms before
this commit).

We consider this change non-breaking and confirmed that (on higher
optimization levels) its performance impact is negligible.