coverage: Separate `derive(Debug)` and exclude from coverage report
On some enums, the default implementation of `core::fmt::Debug` is detected
as non-covered branches (possibly because of "other value" handling) which
should never be covered by a safe program.

To workaround this issue and improve the coverage, this commit separates
`derive(Debug)` from other derivations and excludes them from the
coverage report.

Note that, it seems that "exhaustiveness" of an enum is irrelevant to this
issue and thus this commit adds this workaround for every enums.