tests: Add unit test for `NumericWindows` alone
We had combined tests for relations between `Windows` and `NumericWindows`
but that was a part of the `hash` tests.  This is because the primary
intent was to test whether `block_hash_[12]_windows()` and
`block_hash_[12]_numeric_windows()` have one-to-one correspondence.

This commit adds tests for `NumericWindows` alone (independent to
`FuzzyHashData`).