doc: Adjust feature gate
Rust doctests are built with the standard library (std) with specified
features.  So, just using the vector (`Vec`) doesn't require the example
to require the `alloc` feature.  Likewise, because we have `ToString` trait
in the prelude, we don't need feature gating with `alloc` on the doctests.

This commit reflects that fact and adjusts the doctests *and* the
description in the comments.