generate: Encapsulate internal data of `Generator`
as the separate struct `GeneratorInnerData`.  The reason behind this change
is to provide access to `Copy` and `PartialEq` inside this crate but not
outside.  In other words, we want whatever we do inside the crate
(especially in tests) but we don't want to let users to use them since
copying this struct (implicitly) takes much time and comparing with another
generator makes no sense (in general).