hash::algorithms: Easy-to-maintain parsers
This commit changes those two functions:

1.  `parse_block_size_from_bytes`
2.  `parse_block_hash_from_bytes`

so that the temporary buffer is truncated on-the-fly for easier
maintenance and proper buffer management.

Criterion micro-benchmark showed same class of the performance but some
test apps resulted in up to 15% parser slowdown.  Even in the worst case,
the author considers it's acceptable because we should avoid using the
parser if we need the best performance.

Note that, this change is designed to be simple and short.