Replace "large trigger last hash" sample to an archive
This commit replaces the "large trigger last hash" file to a GNU Tar +
Zstandard archive file (in GNU format).

Since it is now an archive file containing a sparse file, lots of Unix-based
systems can handle the output without consuming much of disk space.  And it
is so small that containing this file in the crate won't cause large issues.

In sync with this, a warning message about zip bomb is weakened
(still, it *can* be a zip bomb though).