nightly: Remove redundant imports
The latest Nightly compiler now generates warnings about "redundant imports".
It happens when the same item is imported twice and can be a problem in
this crate when importing a `std` item for tests *and* the feature `std`
is already enabled.

This commit adds some `#[cfg]` attributes to disable importing related
items when the `std` feature is enabled.