generate: Generate `FNV_TABLE` by code
Through the experiments in `fast-tlsh`, the author observed that
initializing an array with a constant function seems not a bad choice
(though it's a bit uglier than regular Rust code if we stick to the
stable compiler) because it's easy to understand what the initialization
function is doing.

This commit replaces FNV table generation by build script with a constant
initialization function.