hash::block: Index windows (numeric windows with block size)
It packs both 7-length windows in the block hash and its effective
block size.  Due to this property, this is more convenient to use it as
an index (as it contains block size information).

Note that `FuzzyHashData`-side interface is effectively deprecated from the
start because `block_hash_[12]()` functions will start returning basic block
hash proxy objects in the next version.