Implement `Error` either in `std` or `core`
Since the first `error_in_core`-stable Rust compiler, version 1.81 is
released, we are now safe to use `core::error::Error` in the stable
compiler.

But because our MSRV target is currently 1.56, we don't use
`core::error::Error` directly but instead we depend on a macro and the
build script to detect compiler version and switch the implementation,
depending on the configuration.

It also deduplicates two `Error` implementations.