Version 0.3.15 "Prepare for 0.4! (3)"
While the feature itself is unchanged, its internals are being changed to
minimize differences between working version 0.4 branch (despite that
the most of version 0.4 implementation will be an atomic commit, minimizing
the differences is important for fast understandings).

There are also many documentation fixes since the author is now reviewing
the version 0.3 codebase to reimplement version 0.4.

*   Deprecations of some features (see the documentation).
*   Various documentation fixes.
*   Internal changes
    *   Splitted internal hashes inside the generator to make vendoring
        easier (vendoring is an option because they are going to be private
        on the next major release).
    *   Invariants (internally used in the crate) are easier to use.
*   Tidying to minimize differences between this version and version 0.4.
    *   This is in the progress and version 0.3.16 will have
        bigger differences (without making no public differences).