doc: position_array: Add "going to be private" notices
In the next major release, many of internal functionalities
(except small utilities directly related to comparison, currently in the
`FuzzyHashCompareTarget` type) are going to be completely private inside
this crate.

Note: such exceptions will be also moved out from
the `FuzzyHashCompareTarget` type.

This commit adds "going to be private" notices on such types on the
generator and deprecates public `internal_comparison` module.